Adding Auth to tasks enpoint

+++ 01-task-management/src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -15,12 +16,23 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) { }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User
+    ): Promise<Task> {
+        const found = await this.taskRepository.findOne({
+            where: {
+                id,
+                userId: user.id
+            }
+        });
 
         if (!found) {
             throw new NotFoundException(`Task with id ${id} not found`);
@@ -29,20 +41,30 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (!result.affected) {
             throw new NotFoundException(`Task with id ${id} not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
 
         task.status = status;
         await task.save();
